+++ client/src/App.jsx
@@ -15,7 +15,6 @@ class App extends Component {
           <Header />
           <Switch>
             <Route exact path="/" component={Pages.Homepage} />
-            <PrivateRoute exact path="/browse" component={Pages.Browse} />
             <PrivateRoute exact path="/courses/:courseId" component={Pages.Courses} />
             <PrivateRoute exact path="/lessons/:lessonId" component={Pages.Lessons} />
             <PrivateRoute exact path="/profile" component={Pages.Profile} />
