+++ client/src/pages/index.jsx
@@ -4,7 +4,6 @@ import About from './About/About';
 import Profile from './Profile/Profile';
 import Courses from './Courses/Courses';
 import Lessons from './Lessons/Lessons';
-import Browse from './Browse/Browse';
 import Subject from './Subject/Subject';
 import Login from './Login/Login';
 import Signup from './Signup/Signup';
@@ -12,6 +11,6 @@ import Error404 from './404/404';
 import Admin from './Admin/Admin';
 
 export {
-    Homepage, About, Profile, Courses, Lessons, Browse, Subject, 
+    Homepage, About, Profile, Courses, Lessons, Subject, 
     Login, Signup, Error404, Admin
 }
